+++ backend/src/main/java/com/ceiba/fashtoll/searchEngine/SearchTokenRepository.java
@@ -1,6 +1,6 @@
-package com.ceiba.fashtoll.searchEngine.repositories;
+package com.ceiba.fashtoll.searchEngine;
 
-import com.ceiba.fashtoll.searchEngine.indexingComponent.SearchToken;
+import com.ceiba.fashtoll.searchEngine.entities.SearchToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
