+++ backend/src/main/java/com/ceiba/fashtoll/searchEngine/TemplateMethod/ConcreteSearchEngines/SimpleSearchEngine.java
@@ -2,13 +2,12 @@ package com.ceiba.fashtoll.searchEngine.TemplateMethod.ConcreteSearchEngines;
 
 import com.ceiba.fashtoll.searchEngine.TemplateMethod.SearchEngine;
 import com.ceiba.fashtoll.searchEngine.dtos.QueryFilters;
-import com.ceiba.fashtoll.searchEngine.indexingComponent.IndexingComponent;
-import com.ceiba.fashtoll.searchEngine.rankingComponent.RankingComponent;
+import com.ceiba.fashtoll.searchEngine.IndexingComponent;
+import com.ceiba.fashtoll.searchEngine.RankingComponent;
 import com.ceiba.fashtoll.worldModel.product.entities.Product;
 import com.ceiba.fashtoll.worldModel.product.repositories.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 import java.util.List;
