+++ backend/src/main/java/com/ceiba/fashtoll/searchEngine/TemplateMethod/SearchEngine.java
@@ -2,15 +2,14 @@ package com.ceiba.fashtoll.searchEngine.TemplateMethod;
 
 import com.ceiba.fashtoll.searchEngine.dtos.ProductSearchRequest;
 import com.ceiba.fashtoll.searchEngine.dtos.QueryFilters;
-import com.ceiba.fashtoll.searchEngine.indexingComponent.IndexingComponent;
-import com.ceiba.fashtoll.searchEngine.rankingComponent.RankingComponent;
+import com.ceiba.fashtoll.searchEngine.IndexingComponent;
+import com.ceiba.fashtoll.searchEngine.RankingComponent;
 import com.ceiba.fashtoll.utilities.Singleton.Analyzer;
 import com.ceiba.fashtoll.worldModel.product.entities.Product;
 import com.ceiba.fashtoll.worldModel.product.repositories.ProductRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 
 import java.util.ArrayList;
 import java.util.List;
