--- backend/src/main/java/com/ceiba/fashtoll/searchEngine/indexingComponent/FrequenInfo.java
@@ -1,4 +1,0 @@
-package com.ceiba.fashtoll.searchEngine.indexingComponent;
-
-public class FrequenInfo {
-}
