--- backend/src/main/java/com/ceiba/fashtoll/searchEngine/rankingComponent/UserPreferences.java
@@ -1,4 +1,0 @@
-package com.ceiba.fashtoll.searchEngine.rankingComponent;
-
-public class UserPreferences {
-}
